[fix-loading-issues-plugin] fix duplicate pluginName when multiple errors; fix redirect problems with plugins
