+++ bbb-common-web/src/main/java/org/bigbluebutton/api/service/RedirectValidator.java
@@ -1,0 +1,5 @@
+package org.bigbluebutton.api.service;
+
+public interface RedirectValidator {
+    boolean isRedirectValid(String redirectURL);
+}
